requirements.txt
flask
google-cloud-logging==2.0.2
requests
